- Moved the implementation to the ast package so it doesn't conflict with current AST - Very minor syntax adjustments
