+++ jvm/src/test/scala/noInline/org/json4s/ast/package.scala
@@ -1,6 +1,6 @@
-package noInline.org
+package noInline.org.json4s
 
-package object json4s {
+package object ast {
   trait BigDecimalConverter[T]{  def apply(b: BigDecimal): T }
 
   object BigDecimalConverter {
